+++ src/platform/sim/sim/frontend/widgets/Display.h
@@ -12,7 +12,7 @@ class Display : public Widget {
 public:
     typedef std::shared_ptr<Display> Ptr;
 
-    Display(const Vector2i &pos, const Vector2i &size, const Vector2i &resolution, const Color &color = Color(0.8f, 0.9f, 0.f, 1.f)) :
+    Display(const Vector2f &pos, const Vector2f &size, const Vector2i &resolution, const Color &color = Color(0.8f, 0.9f, 0.f, 1.f)) :
         Widget(pos, size),
         _resolution(resolution),
         _color(color),
